Fix bug with one sided walls
I might want to fix this in a better way in the future. Right now it
seems I am wasting memory because I have to replicate the wall data for
every wall I insert. However, a solution where a single wall lookup is
implemented would increase the complexity of the lookup itself. Would it
be worth it?