Make round trip
This technically completes the assignment but there are still some
todos. For example, I want to fine tune the BFS algorithm so that after
finding the best round it runs a second time, but this time it will
count the number of turns and will trim the branches with the most
turns. Another thing that I want to do is createa a buffer for an
optimal route. This is what I called in the code speedrun mode. It will
potentially add more power to the motors to make the fastest time it
can. It is possible that in this state the algorithms are turned off and
it will run purely from memory.