Include depedencies
This is not working as expected. I have no idea why. I have to take a
look at this in the future. The intent is that by changing a header all
the appropriate c files will re-compile. I am not sure how practical it
is that anyway since changes can only affect those translation units
which actually changed. However, one scenario where I can see something
fail is when the signature of a function is updated along with its
definition. The translation units which use the header public interface
may not change and try to use the old signature.